Move date utils to own file
